A1.8.2
----
System
----
-Changed Mob damage to damage calculator
-Added Sector 2 and 4 to manager
-Added Secotr 4 to tp command
-Fixed vote update to players
-potions can now be permeant
-Lowered based portal drone sound
----
Game
------
-New Tier one sector added
-Nerfed Ranger damage scaling from 100% STG + 50% AG to 80% STG + 50% AG
-Added the Glub a T1 ranged enemy
-Buffed Dragon power from 25 to 35 and attack cooldown from 2 to 1.7
-Buffed Dragon evolved power from 40 to 45

+++ Assets/Scripts/Items/Projectiles/Spell Prefab/Base Scripts/Projectile.cs
@@ -137,9 +137,9 @@ public class Projectile : MonoBehaviourPun
                     if (enemy.IsLocal == false)
                     {
                         if (damageType == DamageType.Magic)
-                            enemy.TakeDamge(DamageCaculator.MagicDamage(character, (int)DamageCaculator.Reciver.Player, null, enemy, ProjectileToUse.dmgAmt));
+                            enemy.TakeDamge(DamageCaculator.MagicDamage(character, DamageCaculator.Reciver.Player, null, enemy, ProjectileToUse.dmgAmt));
                         if (damageType == DamageType.Ranged)
-                            enemy.TakeDamge(DamageCaculator.RangedDamage(character, (int)DamageCaculator.Reciver.Player, null, enemy, ProjectileToUse.dmgAmt));
+                            enemy.TakeDamge(DamageCaculator.RangedDamage(character, DamageCaculator.Reciver.Player, null, enemy, ProjectileToUse.dmgAmt));
                         if (statusEffects != null)
                         {
                             foreach (StatusEffectSO effect in statusEffects)
